fix: Updates api key so static map can display (#1692)
* fix: Updates api key so static map can display

Investigation revealed that the static map API key has been set to "YOUR_API_KEY", rendering the live example somewhat useless. Since there's already code on the page to get an API  key, I restructured the call to get the key when the page loads, and pass that value to the static map so that it can actually load.

* Update index.ts

added a ts-ignore for a VIte bug, does not like `env`.