fix: Removes exclusion tags, changes to use DEMO_MAP_ID (#1697)
* fix: Removes exclusion tags, changes to use DEMO_MAP_ID

Problem: The example map had used a custom reduced-POI map ID; a subsequent request was made to obscure the map ID parameter to avoid exposing the custom map ID. The example works when users link to JS Fiddle using the buttons, but if the code is copied and pasted it won't work.

Solution: Get rid of the custom map ID and use DEMO_MAP_ID; remove the exclude tag which hides the needed map ID parameter.

This fix is in response to b/316906523

* Adds type declaration to bounds.extend

Changes ln. 60 to include type declaration.

* Adds type declaration for bounds.extend

Updates ln. 65 to include a type declaration for bounds.extend.